+++ packages/native/src/components/Box/index.ts
@@ -13,9 +13,9 @@ import {
 
 import { BoxStyleProps } from './style/types'
 
-export const Box = styled.View.withConfig({
+export const Box = styled.View.withConfig<BoxStyleProps>({
   shouldForwardProp: (prop) => !['elevation', 'radius'].includes(prop),
-})<BoxStyleProps>`
+})`
   display: flex;
   position: relative;
   border-style: solid;